refactor: Reservation 객체의 id 발급 로직을 Repository로 이동
